fix/ESEU-170 umożliwienie wprowadzania liczb dziesiętnych w pole currentTermPayment

+++ src/groovy/com/eservice/eumowy/command/HirePaymentCommand.groovy
@@ -24,8 +24,8 @@ class HirePaymentCommand implements Serializable{
     String type
     Integer termCount
     Integer ppCount
-    Integer currentTermPayment
-    Integer currentPpPayment
+    BigDecimal currentTermPayment
+    BigDecimal currentPpPayment
     String newTermPayment
     String newPpPayment
     Boolean isChoosen
